Update slapd.conf.erb
The slapd.conf was missing the module section in template when I ran the cookbook.  I have tested in Centos 7.4.1708

Created a wrapper cookbook then populated the default['openldap']['modules'] = %w(back_hdb memberof) and added the memberof.
I ran the cookbook the first time and went into the slapd.conf and noticed modules section was missing.  I traced it back to this template.

Just want to submit a fix for what I think you meant to do in the template.   That is load module name but skip module path.  Let me know if this is wrong.